agregando css

+++ familiares/models.py
@@ -3,6 +3,6 @@ from django.forms import CharField
 
 class Familiares(models.Model):
     name = models.CharField(max_length=50)
-    age = models.FloatField()
+    age = models.IntegerField()
     active = models.BooleanField(default=True)
     date = models.CharField(max_length=50, unique=True)
